Hardcode the symbol expressions
